Created a run_model executable to run models on the user container.  Modified the run_models handler to send a command to execute the run_model executable.  Modified the singleuser dockerfile to copy the run_model.py file to the users file system.  Fixed a bug with the public files naming schema.
