+++ public_models/Brusselator.mdl
