+++ public_models/Vilar_Oscillator.mdl
